Section3-CloudBuild-CICD/cloudrun-ml-models/coupon-recommendations/requirements.txt
Flask>=3.0,<4.0
Werkzeug>=3.0,<4.0
gunicorn>=21.2,<22
pytest>=8,<9
numpy<2
pandas<2.2
pyarrow>=15,<23
google-cloud-bigquery>=3,<4
scikit-learn==1.2.2
xgboost==1.6.2
category_encoders==2.6.3
imbalanced-learn
requests
google-cloud-storage